agent/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
pymongo==4.12.1
pydantic==2.13.0
langgraph==0.3.5
langchain-google-genai==2.0.11
langchain-core==0.3.42
python-dotenv==1.0.1
